// apps/frontend/src/generated/api-contract.ts
export type AddressState = "available" | "reserved" | "assigned" | "observed" | "conflict" | "excluded";
// ...
export interface NewIpAddress {
  routingDomainId: string;
  prefixId: string;
  deviceId?: string | null;
  interfaceId?: string | null;
  address: string;
  dnsName?: string | null;
  state: AddressState;
  source: string;
  description?: string | null;
}

export interface NewIpPrefix {
  routingDomainId: string;
  prefix: string;
  name?: string | null;
  gateway?: string | null;
  vlanId?: number | null;
  description?: string | null;
  purpose?: string | null;
}
// ...
export interface UpdateIpAddress {
  expectedVersion: number;
  deviceId?: string | null;
  interfaceId?: string | null;
  dnsName?: string | null;
  state: AddressState;
  source: string;
  description?: string | null;
}

export interface UpdateIpPrefix {
  expectedVersion: number;
  routingDomainId: string;
  prefix: string;
  name?: string | null;
  gateway?: string | null;
  vlanId?: number | null;
  description?: string | null;
  purpose?: string | null;
}